Update Made to repo

+++ 2012/gb_ubuntu_USN_1386_1.nasl
@@ -1,6 +1,6 @@
 ###############################################################################
 # OpenVAS Vulnerability Test
-# $Id: gb_ubuntu_USN_1386_1.nasl 9352 2018-04-06 07:13:02Z cfischer $
+# $Id: gb_ubuntu_USN_1386_1.nasl 9649 2018-04-27 08:45:50Z cfischer $
 #
 # Ubuntu Update for linux-lts-backport-natty USN-1386-1
 #
@@ -25,7 +25,32 @@
 ###############################################################################
 
 include("revisions-lib.inc");
-tag_insight = "The linux kernel did not properly account for PTE pages when deciding which
+
+
+
+if(description)
+{
+  script_xref(name: "URL" , value: "http://www.ubuntu.com/usn/usn-1386-1/");
+  script_oid("1.3.6.1.4.1.25623.1.0.840920");
+  script_version("$Revision: 9649 $");
+  script_tag(name:"last_modification", value:"$Date: 2018-04-27 10:45:50 +0200 (Fri, 27 Apr 2018) $");
+  script_tag(name:"creation_date", value:"2012-03-07 11:19:56 +0530 (Wed, 07 Mar 2012)");
+  script_cve_id("CVE-2011-2498", "CVE-2011-2518", "CVE-2011-3353", "CVE-2011-4097",
+                "CVE-2011-4622", "CVE-2012-0038", "CVE-2012-0044", "CVE-2012-0207");
+  script_tag(name:"cvss_base", value:"7.8");
+  script_tag(name:"cvss_base_vector", value:"AV:N/AC:L/Au:N/C:N/I:N/A:C");
+  script_xref(name: "USN", value: "1386-1");
+  script_name("Ubuntu Update for linux-lts-backport-natty USN-1386-1");
+
+  script_category(ACT_GATHER_INFO);
+  script_copyright("Copyright (c) 2012 Greenbone Networks GmbH");
+  script_family("Ubuntu Local Security Checks");
+  script_dependencies("gather-package-list.nasl");
+  script_mandatory_keys("ssh/login/ubuntu_linux", "ssh/login/packages", re:"ssh/login/release=UBUNTU10\.04 LTS");
+  script_tag(name : "summary" , value : "Ubuntu Update for Linux kernel vulnerabilities USN-1386-1");
+  script_tag(name : "affected" , value : "linux-lts-backport-natty on Ubuntu 10.04 LTS");
+  script_tag(name : "solution" , value : "Please Install the Updated Packages.");
+  script_tag(name : "insight" , value : "The linux kernel did not properly account for PTE pages when deciding which
   task to kill in out of memory conditions. A local, unprivileged could
   exploit this flaw to cause a denial of service. (CVE-2011-2498)
 
@@ -55,37 +80,7 @@ tag_insight = "The linux kernel did not properly account for PTE pages when deci
   (CVE-2012-0044)
 
   A flaw was found in the linux kernels IPv4 IGMP query processing. A remote
-  attacker could exploit this to cause a denial of service. (CVE-2012-0207)";
-
-tag_summary = "Ubuntu Update for Linux kernel vulnerabilities USN-1386-1";
-tag_affected = "linux-lts-backport-natty on Ubuntu 10.04 LTS";
-tag_solution = "Please Install the Updated Packages.";
-
-
-
-if(description)
-{
-  script_xref(name: "URL" , value: "http://www.ubuntu.com/usn/usn-1386-1/");
-  script_oid("1.3.6.1.4.1.25623.1.0.840920");
-  script_version("$Revision: 9352 $");
-  script_tag(name:"last_modification", value:"$Date: 2018-04-06 09:13:02 +0200 (Fri, 06 Apr 2018) $");
-  script_tag(name:"creation_date", value:"2012-03-07 11:19:56 +0530 (Wed, 07 Mar 2012)");
-  script_cve_id("CVE-2011-2498", "CVE-2011-2518", "CVE-2011-3353", "CVE-2011-4097",
-                "CVE-2011-4622", "CVE-2012-0038", "CVE-2012-0044", "CVE-2012-0207");
-  script_tag(name:"cvss_base", value:"7.8");
-  script_tag(name:"cvss_base_vector", value:"AV:N/AC:L/Au:N/C:N/I:N/A:C");
-  script_xref(name: "USN", value: "1386-1");
-  script_name("Ubuntu Update for linux-lts-backport-natty USN-1386-1");
-
-  script_category(ACT_GATHER_INFO);
-  script_copyright("Copyright (c) 2012 Greenbone Networks GmbH");
-  script_family("Ubuntu Local Security Checks");
-  script_dependencies("gather-package-list.nasl");
-  script_mandatory_keys("ssh/login/ubuntu_linux", "ssh/login/packages");
-  script_tag(name : "summary" , value : tag_summary);
-  script_tag(name : "affected" , value : tag_affected);
-  script_tag(name : "solution" , value : tag_solution);
-  script_tag(name : "insight" , value : tag_insight);
+  attacker could exploit this to cause a denial of service. (CVE-2012-0207)");
   script_tag(name:"qod_type", value:"package");
   script_tag(name:"solution_type", value:"VendorFix");
   exit(0);
@@ -128,6 +123,6 @@ if(release == "UBUNTU10.04 LTS")
     exit(0);
   }
 
-  if (__pkg_match) exit(99); # Not vulnerable.
+  if (__pkg_match) exit(99);
   exit(0);
 }
